+++ tests/Integration/ManufacturerTest.php
@@ -1,0 +1,17 @@
+<?php
+
+
+namespace Tests;
+
+
+class ManufacturerTest extends \Jtl\Connector\IntegrationTests\Integration\ManufacturerTest
+{
+    public function getIgnoreArray()
+    {
+        return [
+            'sort',
+            'urlPath',
+            'websiteUrl'
+        ];
+    }
+}