+++ tests/integration/CategoryTest.php
@@ -1,0 +1,17 @@
+<?php
+
+
+namespace Tests;
+
+
+class CategoryTest extends \ConnectorIntegrationTests\Integration\CategoryTest
+{
+    public function getIgnoreArray()
+    {
+        return [
+            'level',
+            'id',
+            'i18ns'
+        ];
+    }
+}